Add Tesco1,2&3 and TJ Morris Model 1 via data-models
Also, fix folder names
