+++ app/services/parsers/tescos/model2/parser.js
@@ -3,6 +3,7 @@ const ParserModel = require("../../../parser-model");
 
 function parse(packingListJson) {
   const establishmentNumber = packingListJson[2].M ?? null;
+  console.log("establishmentNumber: ", establishmentNumber);
   const packingListContents = packingListJson.slice(2).map((col) => ({
     description: col.F ?? null,
     nature_of_products: null,
@@ -11,6 +12,7 @@ function parse(packingListJson) {
     number_of_packages: col.H ?? null,
     total_net_weight_kg: col.K ?? null,
   }));
+  console.log("packingListContents: ", packingListContents);
 
   return CombineParser.combine(
     establishmentNumber,
