--- test/unit/services/parser-service/asda/model-1/parser-service.test.js
@@ -1,21 +1,0 @@
-const parserService = require("../../../../../../app/services/parser-service");
-const model = require("../../../test-helpers/asda/model1/data-model");
-
-const filename = "packinglist-asda-model1.xls";
-
-describe("matchesAsdaModel1", () => {
-  test("matches valid Asda Model 1 file, calls parser and returns all_required_fields_present as true", () => {
-    const result = parserService.findParser(model.validModel, filename);
-
-    expect(result).toEqual(model.validTestResult);
-  });
-
-  test("matches valid Asda Model 1 file, calls parser, but returns all_required_fields_present as false when cells missing", () => {
-    const result = parserService.findParser(
-      model.invalidModel_MissingColumnCells,
-      filename,
-    );
-
-    expect(result).toEqual(model.invalidTestResult_MissingCells);
-  });
-});
