+++ test/unit/test-helpers/tescos/model2/data-model.js
@@ -1,0 +1,168 @@
+const ParserModel = require("../../../../../app/services/parser-model");
+
+module.exports = {
+  validModel: {
+    Sheet2: [
+      {
+        A: "Item",
+        B: "Product code",
+        C: "Commodity code",
+        D: "Online Check",
+        E: "Meursing code",
+        F: "Description of goods",
+        G: "Country of Origin",
+        H: "No. of pkgs",
+        I: "Type of pkgs",
+        J: "Total Gross Weight",
+        K: "Total Net Weight",
+        L: "Total Line Value",
+        M: "GB Establishment RMS Number",
+      },
+      {},
+      {
+        A: "1",
+        B: "SKU1944",
+        C: "2005995090",
+        F: "TF R/Bow Tom with Blsac Glze 340g x4",
+        G: "Great Britain",
+        H: "4",
+        I: "Cases",
+        J: "16",
+        K: "9.312",
+        L: "31.04",
+        M: "RMS-GB-000015-009",
+      },
+      {
+        A: "2",
+        B: "SKU1938",
+        C: "2005995090",
+        F: "TF 300g Roasting Vegetables x8",
+        G: "Great Britain",
+        H: "4",
+        I: "Cases",
+        J: "32",
+        K: "16.144",
+        L: "64.32",
+        M: "RMS-GB-000015-009",
+      },
+    ],
+  },
+  invalidModel_MissingColumnCells: {
+    Sheet2: [
+      {
+        A: "Item",
+        B: "Product code",
+        C: "Commodity code",
+        D: "Online Check",
+        E: "Meursing code",
+        F: "Description of goods",
+        G: "Country of Origin",
+        H: "No. of pkgs",
+        I: "Type of pkgs",
+        J: "Total Gross Weight",
+        K: "Total Net Weight",
+        L: "Total Line Value",
+        M: "GB Establishment RMS Number",
+      },
+      {},
+      {
+        A: "1",
+        B: "SKU1944",
+        C: "2005995090",
+        F: "TF R/Bow Tom with Blsac Glze 340g x4",
+        G: "Great Britain",
+        H: "4",
+        I: "Cases",
+        J: "16",
+        K: null,
+        L: "31.04",
+        M: "RMS-GB-000015-009",
+      },
+      {
+        A: "2",
+        B: "SKU1938",
+        C: null,
+        F: "TF 300g Roasting Vegetables x8",
+        G: "Great Britain",
+        H: "4",
+        I: "Cases",
+        J: "32",
+        K: "16.144",
+        L: "64.32",
+        M: "RMS-GB-000015-009",
+      },
+    ],
+  },
+  emptyModel: {
+    Sheet2: [
+      {
+        A: "Item",
+        B: "Product code",
+        C: "Commodity code",
+        D: "Online Check",
+        E: "Meursing code",
+        F: "Description of goods",
+        G: "Country of Origin",
+        H: "No. of pkgs",
+        I: "Type of pkgs",
+        J: "Total Gross Weight",
+        K: "Total Net Weight",
+        L: "Total Line Value",
+        M: "GB Establishment RMS Number",
+      },
+      {
+        M: null,
+      },
+    ],
+  },
+  validTestResult: {
+    business_checks: {
+      all_required_fields_present: true,
+    },
+    items: [
+      {
+        commodity_code: "2005995090",
+        description: "TF R/Bow Tom with Blsac Glze 340g x4",
+        nature_of_products: null,
+        number_of_packages: "4",
+        total_net_weight_kg: "9.312",
+        type_of_treatment: null,
+      },
+      {
+        commodity_code: "2005995090",
+        description: "TF 300g Roasting Vegetables x8",
+        nature_of_products: null,
+        number_of_packages: "4",
+        total_net_weight_kg: "16.144",
+        type_of_treatment: null,
+      },
+    ],
+    registration_approval_number: "RMS-GB-000015-009",
+    parserModel: ParserModel.TESCO2,
+  },
+  invalidTestResult_MissingCells: {
+    business_checks: {
+      all_required_fields_present: false,
+    },
+    items: [
+      {
+        commodity_code: "2005995090",
+        description: "TF R/Bow Tom with Blsac Glze 340g x4",
+        nature_of_products: null,
+        number_of_packages: "4",
+        total_net_weight_kg: null,
+        type_of_treatment: null,
+      },
+      {
+        commodity_code: null,
+        description: "TF 300g Roasting Vegetables x8",
+        nature_of_products: null,
+        number_of_packages: "4",
+        total_net_weight_kg: "16.144",
+        type_of_treatment: null,
+      },
+    ],
+    registration_approval_number: "RMS-GB-000015-009",
+    parserModel: ParserModel.TESCO2,
+  },
+};
